+++ lib/devices/device_factory.dart
@@ -1,10 +1,20 @@
 import '../export/fit/fit_manufacturer.dart';
 import '../utils/constants.dart';
 import 'device_descriptors/cross_trainer_device_descriptor.dart';
+import 'device_descriptors/concept2_rower.dart';
 import 'device_descriptors/cycling_power_meter_descriptor.dart';
 import 'device_descriptors/cycling_speed_and_cadence_descriptor.dart';
+import 'device_descriptors/device_descriptor.dart';
 import 'device_descriptors/indoor_bike_device_descriptor.dart';
+import 'device_descriptors/matrix_bike_descriptor.dart';
+import 'device_descriptors/matrix_treadmill_descriptor.dart';
+import 'device_descriptors/mr_captain_descriptor.dart';
+import 'device_descriptors/npe_runn_treadmill.dart';
+import 'device_descriptors/paddling_speed_and_cadence_descriptor.dart';
+import 'device_descriptors/precor_spinner_chrono_power.dart';
 import 'device_descriptors/rower_device_descriptor.dart';
+import 'device_descriptors/schwinn_ac_performance_plus.dart';
+import 'device_descriptors/schwinn_x70.dart';
 import 'device_descriptors/treadmill_device_descriptor.dart';
 import 'device_fourcc.dart';
 
@@ -14,7 +24,6 @@ class DeviceFactory {
       fourCC: schwinnICBikeFourCC,
       vendorName: "Nautilus, Inc",
       modelName: "Schwinn IC4/IC8",
-      namePrefixes: ["IC Bike"],
       manufacturerNamePart: "Nautilus",
       manufacturerFitId: nautilusFitId,
       model: "IC BIKE",
@@ -27,7 +36,6 @@ class DeviceFactory {
       fourCC: bowflexC7BikeFourCC,
       vendorName: "Nautilus Inc.",
       modelName: "Bowflex C7",
-      namePrefixes: ["C7-"],
       manufacturerNamePart: "Nautilus",
       manufacturerFitId: nautilusFitId,
       model: "Bowflex C7",
@@ -40,7 +48,6 @@ class DeviceFactory {
       fourCC: schwinnUprightBikeFourCC,
       vendorName: "Nautilus, Inc",
       modelName: "Schwinn 230/510",
-      namePrefixes: ["SCH130", "SCH230", "SCH510"],
       manufacturerNamePart: "Nautilus",
       manufacturerFitId: nautilusFitId,
       model: "SCH BIKE",
@@ -52,7 +59,6 @@ class DeviceFactory {
       fourCC: stagesSB20FourCC,
       vendorName: "Stages Cycling",
       modelName: "SB20",
-      namePrefixes: ["Stages Bike"],
       manufacturerNamePart: "Stages",
       manufacturerFitId: stagesCyclingFitId,
       model: "SB20",
@@ -64,7 +70,6 @@ class DeviceFactory {
       fourCC: yesoulS3FourCC,
       vendorName: "Yesoul",
       modelName: "S3",
-      namePrefixes: ["Yesoul"],
       manufacturerNamePart: "Yesoul",
       manufacturerFitId: stravaFitId,
       model: "S3",
@@ -73,11 +78,10 @@ class DeviceFactory {
 
   static RowerDeviceDescriptor getKayaPro() {
     return RowerDeviceDescriptor(
-      defaultSport: ActivityType.kayaking,
+      sport: deviceSportDescriptors[kayakProGenesisPortFourCC]!.defaultSport,
       fourCC: kayakProGenesisPortFourCC,
       vendorName: "KayakPro",
       modelName: "KayakPro Compact",
-      namePrefixes: ["KayakPro", "KP"],
       manufacturerNamePart: "North Pole Engineering",
       manufacturerFitId: northPoleEngineeringFitId,
       model: "64",
@@ -89,7 +93,6 @@ class DeviceFactory {
       fourCC: genericFTMSTreadmillFourCC,
       vendorName: "Unknown",
       modelName: "Generic Treadmill",
-      namePrefixes: ["FTMS Treadmill"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Treadmill",
@@ -101,7 +104,6 @@ class DeviceFactory {
       fourCC: genericFTMSBikeFourCC,
       vendorName: "Unknown",
       modelName: "Generic Indoor Bike",
-      namePrefixes: ["FTMS Bike"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Indoor Bike",
@@ -110,12 +112,11 @@ class DeviceFactory {
 
   static RowerDeviceDescriptor getGenericFTMSKayaker() {
     return RowerDeviceDescriptor(
-      defaultSport: ActivityType.kayaking,
-      isMultiSport: false,
+      sport: deviceSportDescriptors[genericFTMSKayakFourCC]!.defaultSport,
+      isMultiSport: deviceSportDescriptors[genericFTMSKayakFourCC]!.isMultiSport,
       fourCC: genericFTMSKayakFourCC,
       vendorName: "Unknown",
       modelName: "Generic Kayak Ergometer",
-      namePrefixes: ["FTMS Kayak"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Kayak Ergometer",
@@ -124,12 +125,11 @@ class DeviceFactory {
 
   static RowerDeviceDescriptor getGenericFTMSCanoeer() {
     return RowerDeviceDescriptor(
-      defaultSport: ActivityType.canoeing,
+      sport: deviceSportDescriptors[genericFTMSCanoeFourCC]!.defaultSport,
       isMultiSport: false,
       fourCC: genericFTMSCanoeFourCC,
       vendorName: "Unknown",
       modelName: "Generic Canoe Ergometer",
-      namePrefixes: ["FTMS Canoe"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Canoe Ergometer",
@@ -138,12 +138,11 @@ class DeviceFactory {
 
   static RowerDeviceDescriptor getGenericFTMSRower() {
     return RowerDeviceDescriptor(
-      defaultSport: ActivityType.rowing,
-      isMultiSport: false,
+      sport: deviceSportDescriptors[genericFTMSRowerFourCC]!.defaultSport,
+      isMultiSport: deviceSportDescriptors[genericFTMSRowerFourCC]!.isMultiSport,
       fourCC: genericFTMSRowerFourCC,
       vendorName: "Unknown",
       modelName: "Generic Rower Ergometer",
-      namePrefixes: ["FTMS Rower"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Rower Ergometer",
@@ -152,12 +151,11 @@ class DeviceFactory {
 
   static RowerDeviceDescriptor getGenericFTMSSwimmer() {
     return RowerDeviceDescriptor(
-      defaultSport: ActivityType.swim,
-      isMultiSport: false,
+      sport: deviceSportDescriptors[genericFTMSSwimFourCC]!.defaultSport,
+      isMultiSport: deviceSportDescriptors[genericFTMSSwimFourCC]!.isMultiSport,
       fourCC: genericFTMSSwimFourCC,
       vendorName: "Unknown",
       modelName: "Generic Swim Ergometer",
-      namePrefixes: ["FTMS Swim"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Swim Ergometer",
@@ -167,12 +165,11 @@ class DeviceFactory {
   // Delete this?
   static RowerDeviceDescriptor getGenericFTMSElliptical() {
     return RowerDeviceDescriptor(
-      defaultSport: ActivityType.elliptical,
-      isMultiSport: false,
+      sport: deviceSportDescriptors[genericFTMSEllipticalFourCC]!.defaultSport,
+      isMultiSport: deviceSportDescriptors[genericFTMSEllipticalFourCC]!.isMultiSport,
       fourCC: genericFTMSEllipticalFourCC,
       vendorName: "Unknown",
       modelName: "Generic Cross Elliptical",
-      namePrefixes: ["FTMS Elliptical"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Elliptical",
@@ -184,7 +181,6 @@ class DeviceFactory {
       fourCC: genericFTMSCrossTrainerFourCC,
       vendorName: "Unknown",
       modelName: "Generic Cross Trainer",
-      namePrefixes: ["FTMS Cross Trainer"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Generic Cross Trainer",
@@ -196,7 +192,6 @@ class DeviceFactory {
       fourCC: powerMeterBasedBikeFourCC,
       vendorName: "Unknown",
       modelName: "Power Meter Based Bike",
-      namePrefixes: ["Stages IC"],
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
       model: "Power Meter Based Bike",
@@ -205,13 +200,114 @@ class DeviceFactory {
 
   static CyclingSpeedAndCadenceDescriptor getCSCBasedBike() {
     return CyclingSpeedAndCadenceDescriptor(
-      fourCC: powerMeterBasedBikeFourCC,
+      fourCC: cscSensorBasedBikeFourCC,
       vendorName: "Unknown",
-      modelName: "Speed and Cadence Sensor Based Bike",
-      namePrefixes: ["N/A"],
+      modelName: "Speed and Cadence Sensor Bike",
       manufacturerNamePart: "Unknown",
       manufacturerFitId: stravaFitId,
-      model: "Speed and Cadence Sensor Based Bike",
+      model: "Speed and Cadence Sensor Bike",
     );
+  }
+
+  static PaddlingSpeedAndCadenceDescriptor getCSCBasedPaddler() {
+    return PaddlingSpeedAndCadenceDescriptor(
+      fourCC: cscSensorBasedPaddleFourCC,
+      vendorName: "Old Danube",
+      modelName: "Old Danube",
+      manufacturerNamePart: "Unknown",
+      manufacturerFitId: stravaFitId,
+      model: "Old Danube",
+    );
+  }
+
+  static DeviceDescriptor getDescriptorForFourCC(String fourCC) {
+    switch (fourCC) {
+      case precorSpinnerChronoPowerFourCC:
+        return PrecorSpinnerChronoPower();
+      case schwinnICBikeFourCC:
+        return DeviceFactory.getSchwinnIcBike();
+      case bowflexC7BikeFourCC:
+        return DeviceFactory.getBowflexC7();
+      case schwinnUprightBikeFourCC:
+        return DeviceFactory.getSchwinnUprightBike();
+      case schwinnX70BikeFourCC:
+        return SchwinnX70();
+      case stagesSB20FourCC:
+        return DeviceFactory.getStagesSB20();
+      case yesoulS3FourCC:
+        return DeviceFactory.getYesoulS3();
+      case schwinnACPerfPlusFourCC:
+        return SchwinnACPerformancePlus();
+      case matrixBikeFourCC:
+        return MatrixBikeDescriptor();
+      case kayakProGenesisPortFourCC:
+        return DeviceFactory.getKayaPro();
+      case mrCaptainRowerFourCC:
+        return MrCaptainDescriptor();
+      case npeRunnFourCC:
+        return NpeRunnTreadmill();
+      case matrixTreadmillFourCC:
+        return MatrixTreadmillDescriptor();
+      case genericFTMSTreadmillFourCC:
+        return DeviceFactory.getGenericFTMSTreadmill();
+      case genericFTMSBikeFourCC:
+        return DeviceFactory.getGenericFTMSBike();
+      case genericFTMSKayakFourCC:
+        return DeviceFactory.getGenericFTMSKayaker();
+      case genericFTMSCanoeFourCC:
+        return DeviceFactory.getGenericFTMSCanoeer();
+      case genericFTMSRowerFourCC:
+        return DeviceFactory.getGenericFTMSRower();
+      case genericFTMSSwimFourCC:
+        return DeviceFactory.getGenericFTMSSwimmer();
+      // Delete this?
+      case genericFTMSEllipticalFourCC:
+        return DeviceFactory.getGenericFTMSElliptical();
+      case genericFTMSCrossTrainerFourCC:
+        return DeviceFactory.getGenericFTMSCrossTrainer();
+      case powerMeterBasedBikeFourCC:
+        return DeviceFactory.getPowerMeterBasedBike();
+      case cscSensorBasedBikeFourCC:
+        return DeviceFactory.getCSCBasedBike();
+      case cscSensorBasedPaddleFourCC:
+        return DeviceFactory.getCSCBasedPaddler();
+      case concept2RowerFourCC:
+        return Concept2Rower();
+    }
+
+    return DeviceFactory.getGenericFTMSBike();
+  }
+
+  static DeviceDescriptor genericDescriptorForSport(String sport) {
+    String fourCC = genericFTMSBikeFourCC;
+    switch (sport) {
+      case ActivityType.ride:
+        fourCC = genericFTMSBikeFourCC;
+        break;
+      case ActivityType.run:
+        fourCC = genericFTMSTreadmillFourCC;
+        break;
+      case ActivityType.kayaking:
+        fourCC = genericFTMSKayakFourCC;
+        break;
+      case ActivityType.canoeing:
+        fourCC = genericFTMSCanoeFourCC;
+        break;
+      case ActivityType.rowing:
+        fourCC = genericFTMSRowerFourCC;
+        break;
+      case ActivityType.swim:
+        fourCC = genericFTMSSwimFourCC;
+        break;
+      case ActivityType.elliptical:
+        fourCC = genericFTMSCrossTrainerFourCC;
+        break;
+    }
+
+    return DeviceFactory.getDescriptorForFourCC(fourCC);
+  }
+
+  static List<DeviceDescriptor> allDescriptors() {
+    return [for (var fourCC in allFourCC) DeviceFactory.getDescriptorForFourCC(fourCC)];
   }
 }