+++ test/get_fit_manufacturer_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:track_my_indoor_exercise/devices/device_map.dart';
+import 'package:track_my_indoor_exercise/devices/device_factory.dart';
 import 'package:track_my_indoor_exercise/export/fit/fit_manufacturer.dart';
 
 class TestData {
@@ -11,9 +11,9 @@ class TestData {
 
 void main() {
   group('getFitManufacturer test', () {
-    deviceMap.forEach((fourCC, deviceDescriptor) {
+    DeviceFactory.allDescriptors().forEach((deviceDescriptor) {
       test(
-          "$fourCC (${deviceDescriptor.manufacturerNamePart}) -> ${deviceDescriptor.manufacturerFitId}",
+          "${deviceDescriptor.fourCC} (${deviceDescriptor.manufacturerNamePart}) -> ${deviceDescriptor.manufacturerFitId}",
           () async {
         expect(getFitManufacturer(deviceDescriptor.manufacturerNamePart),
             deviceDescriptor.manufacturerFitId);