+++ test/yesoul_s3_test.dart
@@ -15,7 +15,7 @@ void main() {
   test('Yesoul S3 constructor tests', () async {
     final bike = DeviceFactory.getYesoulS3();
 
-    expect(bike.defaultSport, ActivityType.ride);
+    expect(bike.sport, ActivityType.ride);
     expect(bike.fourCC, yesoulS3FourCC);
     expect(bike.isMultiSport, false);
   });